// src/app/kimios-client-api/model/dMEntity.ts
/**
 * Kimios API
 * Kimios Sample API
 *
 * OpenAPI spec version: 1.2.2-SNAPSHOT - 73e8d1925ff0716b0a39dff428b76dec7209db16
 * Contact: documentation@kimios.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MetaValue } from './metaValue';


export interface DMEntity { 
    uid?: number;
    type?: number;
    name?: string;
    creationDate?: Date;
    updateDate?: Date;
    owner?: string;
    ownerSource?: string;
    path?: string;
    addonDatas?: string;
    metaDatas?: { [key: string]: MetaValue; };
}