// src/app/kimios-client-api/model/metaValue.ts
/**
 * Kimios API
 * Kimios Sample API
 *
 * OpenAPI spec version: 1.2.2-SNAPSHOT - 73e8d1925ff0716b0a39dff428b76dec7209db16
 * Contact: documentation@kimios.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Meta } from './meta';


export interface MetaValue { 
    documentVersionId?: number;
    metaId?: number;
    value?: any;
    meta?: Meta;
}